calendar/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
starlette
google-auth-oauthlib
google-api-python-client
pydantic
firebase-admin
